* https://github.com/icanzilb/EventBlankApp/issues/27
